. gencpp.cpp: fixed a bug builtin-rt client with monitoring support cannot start because libregfs.so was not loaded. . buildall-deb.sh, buildall-rasp.sh makerpcf-rasp.sh: fixed the bugs in makerpcf-rasp.sh.

+++ tools/buildall-rasp.sh
@@ -19,8 +19,18 @@ echo install nodejs 18
 if ! curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.3/install.sh | bash; then exit 1; fi
 if [[ -s $HOME/.nvm/nvm.sh ]]; then source $HOME/.nvm/nvm.sh;nvm install 18;fi
 
-${SUDO} apt-get -y install sip-tools || apt-get -y install sip-dev python3-sip python3-sip-dev || true
-${SUDO} apt-get -y install libjsoncpp1 || apt-get -y install libjsoncpp25 || apt-get -y install libjsoncpp24
+${SUDO} apt-get -y install sip-tools || ${SUDO} apt-get -y install sip-dev python3-sip python3-sip-dev || true
+
+for i in 1 4 5 6; do
+if ${SUDO} apt-get -y install libjsoncpp$i; then 
+    break;
+fi
+done
+if ! dpkg -l libjsoncpp$i > /dev/null; then
+    echo Error unable to install libjsoncpp
+    exit 1
+fi
+
 ${SUDO} apt-get -y install git devscripts debhelper expect screen vim
 ${SUDO} apt-get -y install python3-wheel python3-numpy || pip3 install wheel numpy
